fixed linting errors for the front page
