+++ components/LandingPage/components/index.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import LandingPage from './LandingPage';
 
 export default LandingPage;
