+++ components/Layout/components/index.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import Head from './Head';
 
-export { Head };
+export default Head;
