+++ package.json
@@ -5,7 +5,7 @@
 	"main": "index.js",
 	"scripts": {
 		"test": "jest",
-		"lint": "eslint pages",
+		"lint": "eslint pages components --fix",
 		"dev": "next",
 		"build": "next build",
 		"start": "next start"
@@ -42,6 +42,7 @@
 		"@babel/core": "^7.4.3",
 		"@zeit/next-css": "^1.0.1",
 		"antd": "^3.16.3",
+		"auth0-js": "^9.10.2",
 		"babel-jest": "^24.7.1",
 		"dotenv": "^7.0.0",
 		"express": "^4.16.4",
